chemistryModel, combustionModel: Changed the thermodynamics model argument to const
The chemistryModel and combustionModel do not change the thermodynamics directly
and should not require non-const access to it.  In order to change the
thermodynamics model argument and stored references to const the specie "active"
flags in TDAC have been changed to mutable as this is not a direct change in the
thermodynamic state but a set of switches which allow the state to change
differently during the next thermodynamics update.