topoSet: Added planeToFaceZone
This is a topoSetSource which selects faces based on the adjacent cell
centres spanning a given plane. The plane is defined by a point and
normal vector.

Additionally, an include entry can be specified. When omitted or set to
"all", then all faces that meet the criteria are included in the set. When
set to "closest", just the faces that belong to the closest contiguous
region to the plane point are included. This latter setting is useful when
defining face zones through channels on which the flow rate is to be
computed, as it keeps the set local to a single channel.

An example usage (in system/topoSetDict) is as follows:

    actions
    (
        {
            name    f0;
            type    faceZoneSet;
            action  new;
            source  planeToFaceZone;
            sourceInfo
            {
                point   (0 0 4);
                normal  (1 0 0.2);
                include closest;
            }
        }
    );

This would then allow the flow rate through the created face zone to be
accurately reported by the following command:

    postProcess -func "flowRateFaceZone(name=f0,field=phi)"