Added entrainmentPressure and prghEntrainmentPressure conditions
These provide pressure boundary conditions suitable for use on
boundaries where the flow may reverse or on which the inlet or outlet
state is not known or well defined. The condition takes the following
form:

    p = p0 + 0.5*Un*mag(Un)

In the case of exactly normal inlet velocity, this condition sets the
same pressure as the totalPressure condition. The pressure that is set
increases with increasing outlet velocity and decreases with increasing
inlet velocity. This makes it self-limiting and extremely stable in a
number of configurations which were not easily simulated previously. The
condition also does varies smoothly as the flux reverses, which also
aids stability.

The controls of this boundary condition are exactly the same as for the
totalPressure condition. Note, however, that "p0" is not necessarily the
total pressure any more. It is, in general, a reference pressure.

An example usage is as follows:

    sides
    {
        type    entrainmentPressure;
        p0      100;
    }